add API get,put,delete

+++ src/main/java/ctu/edu/blogAPI/service/UserService.java
@@ -2,13 +2,15 @@ package ctu.edu.blogAPI.service;
 
 import java.util.List;
 
+import javax.management.RuntimeErrorException;
+
 import org.springframework.stereotype.Service;
 import lombok.RequiredArgsConstructor;
 import lombok.AccessLevel;
 import lombok.experimental.FieldDefaults;
-
 import ctu.edu.blogAPI.dto.request.UserCreationRequest;
-import ctu.edu.blogAPI.dto.request.UserUpdateRequest;   // <-- nhớ import
+import ctu.edu.blogAPI.dto.request.UserUpdateRequest; // <-- nhớ import
+import ctu.edu.blogAPI.dto.response.UserRespone;
 import ctu.edu.blogAPI.entities.User;
 import ctu.edu.blogAPI.mapper.UserMapper;
 import ctu.edu.blogAPI.repository.UserRepository;
@@ -18,8 +20,8 @@ import ctu.edu.blogAPI.repository.UserRepository;
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 public class UserService {
 
-    UserRepository userRepository;   // final -> được inject qua constructor
-    UserMapper userMapper;           // final -> được inject qua constructor
+    UserRepository userRepository; // final -> được inject qua constructor
+    UserMapper userMapper; // final -> được inject qua constructor
 
     public User createUser(UserCreationRequest request) {
         if (userRepository.existsByUsername(request.getUsername()))
@@ -33,19 +35,23 @@ public class UserService {
     }
 
     // lấy user theo id (bạn đang gọi hàm này ở updateUser)
-    public User getUser(String id) {
-        return userRepository.findById(id)
-            .orElseThrow(() -> new RuntimeException("User not found"));
+    public UserRespone getUser(String id) {
+        return userMapper.toUserRespone(userRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("User not found")));
     }
 
     // update theo id (dùng MapStruct để bỏ qua null)
-    public User updateUser(String userId, UserUpdateRequest request){
-        User user = getUser(userId);
-        userMapper.updateUser(user, request);   // chỉ ghi đè field != null
-        return userRepository.save(user);
+    public UserRespone updateUser(String userId, UserUpdateRequest request) {
+        User user = userRepository.findAllById(userId)
+                .orElseThrow(() -> new RuntimeErrorException(null, "User not found")); // chỉ ghi đè field != null
+        userMapper.updateUser(user, request);
+        return userMapper.toUserRespone(userRepository.save(user));
     }
 
-    public void deleteUser(String userId){
+    public void deleteUser(String userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new IllegalArgumentException("User not found"); // hoặc NotFoundException tự định nghĩa
+        }
         userRepository.deleteById(userId);
     }
 }
